requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
passlib
requests
python-dotenv
pydantic-sqlalchemy
pyjwt
python-multipart
python-jose
pypdf
langchain-text-splitters
langchain-openai
langchain[mistralai]
langchain-core
langgraph
langchain-community
streamlit